moved the window module into pymodaq_gui
